Add support for GitHub Enterprise instances
We need to make it possible for the user to specify a different GitHub
endpoint in the case that they're pointing it to a GitHub Enterprise
instance.

Because GitHub Enterprise uses a different domain and URL layout than
GitHub.com, we need a logic to specify the API endpoint.

Retain defaults as `github.com`, but allows overriding in user's
`.vimrc`.

Closes #11.